Remove not required unit tests files
