+++ test/experiment_handler_test.py
@@ -1,19 +1,32 @@
 import os
 import subprocess
 import unittest
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, mock_open, patch
 
 import pytest
-from src.prompt_flow_experiment import PromptFlowExperiment
+from src.experiment_handler import ExperimentHandler, PromptFlowExperiment
 
 
-class TestPromptFlowExperiment:
-    @pytest.fixture
-    @patch.object(PromptFlowExperiment, '_run_command')
-    def experiment(self, mock_run_command):
-        mock_run_command.return_value = 0 
-        return PromptFlowExperiment(name="test-experiment")
-    
+class TestExperimentHandler:
+    @patch('src.experiment_handler.ExperimentHandler._run_command')
+    def test_check_and_connect_virtual_env_not_connected(self, mock_run_command):
+        ExperimentHandler()
+        mock_run_command.assert_called_once_with('source .venv/bin/activate')
+
+    @patch('os.environ', {'VIRTUAL_ENV': '/path/to/venv'})
+    @patch('src.experiment_handler.ExperimentHandler._run_command')
+    def test_check_and_connect_virtual_env_connected(self, mock_run_command):
+        ExperimentHandler()
+        mock_run_command.assert_not_called()
+
+    @patch('os.path.exists')
+    @patch('builtins.open', new_callable=mock_open, read_data="VARNAME=value")
+    @patch('src.experiment_handler.ExperimentHandler._run_command')
+    def test_read_and_set_env_vars(self, mock_exists, mock_open, mock_run_command):
+        mock_exists.return_value = True
+        ExperimentHandler()
+        mock_open.assert_called_once_with('.env')
+
     @patch.object(PromptFlowExperiment, '_run_command')
     @patch.object(PromptFlowExperiment, 'create_documentation')
     def test_create(self, mock_create_documentation, mock_run_command):
@@ -25,20 +38,23 @@ class TestPromptFlowExperiment:
         mock_create_documentation.assert_called_once()
 
     @patch('subprocess.Popen')
-    def test_run_command(self, mock_popen, experiment):
+    @patch('os.environ', {'VIRTUAL_ENV': '/path/to/venv'})
+    def test_run_command(self, mock_popen):
         mock_process = MagicMock()
         mock_process.communicate.return_value = (b'', b'')
         mock_process.returncode = 0
         mock_popen.return_value = mock_process
 
         command = 'echo hello'
-        return_code = experiment._run_command(command)
+        experiment_handler = ExperimentHandler()
+        return_code = experiment_handler._run_command(command)
 
         mock_popen.assert_called_once_with(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=os.environ.copy())
         assert return_code == 0
 
-    @patch.object(PromptFlowExperiment, '_run_command')
-    def test_create_resources(self, mock_run_command, experiment):
+    @patch('src.experiment_handler.PromptFlowExperiment._run_command')
+    def test_create_resources(self, mock_run_command):
+        experiment = PromptFlowExperiment("test-experiment")    
         mock_run_command.return_value = 0
 
         experiment.create_resources()
@@ -48,7 +64,8 @@ class TestPromptFlowExperiment:
 
     @patch('shutil.copyfile')
     @patch('builtins.open', new_callable=unittest.mock.mock_open, read_data="This is a template for {{name}}")
-    def test_create_documentation(self, mock_open, mock_copyfile, experiment):
+    def test_create_documentation(self, mock_open, mock_copyfile):
+        experiment = PromptFlowExperiment("test-experiment")    
         experiment.create_documentation()
 
         mock_copyfile.assert_called_once_with('./src/artefacts/TEMPLATE-README.md', f"./app/flow/{experiment.name}/README.md")
@@ -58,7 +75,8 @@ class TestPromptFlowExperiment:
         handle.write.assert_called_once_with('This is a template for test-experiment')
 
     @patch('subprocess.Popen')
-    def test_run_command_failure(self, mock_popen, experiment):
+    def test_run_command_failure(self, mock_popen):
+        experiment = PromptFlowExperiment("test-experiment")    
         mock_process = MagicMock()
         mock_process.communicate.return_value = (b'', b'')
         mock_process.returncode = 1
